test code for set Angle
This is the test code to find the shortest distance between current Angle and target Angle. This code is not tested on a robot. To use this method, the output & input range should of the PID Controller be 360 to  -360 instead of the 180 to -180 which was described in the ether's swerve drive paper